[DQMOffline/Trigger] moved TriggerDQMBase(+others) back to plugins/
