+++ DQMOffline/Trigger/plugins/METDQM.h
@@ -1,7 +1,7 @@
 #ifndef DQMOffline_Trigger_METDQM_H
 #define DQMOffline_Trigger_METDQM_H
 
-#include "DQMOffline/Trigger/interface/TriggerDQMBase.h"
+#include "DQMOffline/Trigger/plugins/TriggerDQMBase.h"
 #include "FWCore/ParameterSet/interface/ParameterSetDescription.h"
 #include "DQMServices/Core/interface/DQMStore.h"
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
